requirements.txt
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.12
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.4.0
idna==3.3
importlib-metadata==4.8.2
pipenv==2021.11.23
platformdirs==2.4.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.10.0
virtualenv-clone==0.5.7
zipp==3.6.0
